Update README.md
# Disney-Hotstar-Clone
In this project, I'm going to show how to clone the popular streaming service Disney + Hotstar clone using nothing but HTML, CSS, and JavaScript. Here is how to create the interface, how to populate it with content, and how to add some basic functionality. 
We will use the following technologies:

- Html
- CSS
- Javascript